I guess this makes sense. also adding sc-im to nix :)

+++ backend/src/ae8900/models/websocket.py
@@ -36,7 +36,7 @@ class PayloadField(BaseModel):
     """
 
     name: str
-    value: int | float
+    value: float | int
     units: str | None
